+++ src/termia/config_actions.py
@@ -104,6 +104,7 @@ class ConfigActionsMixin:
         self.store.data.groups = []
         self.store.data.servers = []
         self.store.data.local_terminals = []
+        self.store.data.workspaces = []
         self.store.save_connections()
         self.selected = None
         self.refresh_list()
@@ -132,6 +133,7 @@ class ConfigActionsMixin:
                     self.store.data.local_terminals,
                     self.store.data.app.connection_storage_mode,
                     self.store.master_password,
+                    self.store.data.workspaces,
                 )
             else:
                 self.store.save_connections()